dialog: Use size_t for length variables (#9572)
Shuts up a warning about size differences.

+++ src/dialog/SDL_dialog_utils.c
@@ -33,7 +33,7 @@ char *convert_filters(const SDL_DialogFileFilter *filters, NameTransform ntf,
     char *new_combined;
     char *converted;
     const char *terminator;
-    int new_length;
+    size_t new_length;
 
     combined = SDL_strdup(prefix);
 
@@ -82,7 +82,7 @@ char *convert_filter(const SDL_DialogFileFilter filter, NameTransform ntf,
 {
     char *converted;
     char *name_filtered;
-    int total_length;
+    size_t total_length;
     char *list;
 
     list = convert_ext_list(filter.pattern, ext_prefix, ext_separator,
@@ -131,7 +131,7 @@ char *convert_ext_list(const char *list, const char *prefix,
 {
     char *converted;
     int semicolons;
-    int total_length;
+    size_t total_length;
 
     semicolons = 0;
